// src/lib/interfaces.ts
export type CustomImageDataFromBing = {
  name: string;
  contentUrl: string;
  hostPageUrl: string;
  hostPageDisplayUrl: string;
  encodingFormat: string;
  contentSize: string;
  width: number;
  height: number;
  accentColor: string;
  imageId: string;
};

export type CustomImageDataFromGoogleSerpapi = {
  position: number;
  thumbnail: string;
  related_content_id: string;
  serpapi_related_content_link: string;
  source: string;
  title: string;
  link: string;
  original: string;
  original_width: number;
  original_height: number;
  is_product: boolean;
};